python_pipeline/pyproject.toml
[project]
name = "sec-filing-pipeline"
version = "0.2.0"
description = "SEC 10-K filing entity extraction pipeline with KuzuDB"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "kuzu>=0.11.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
vertex = [
    "google-cloud-aiplatform>=1.60",
    "google-cloud-storage>=2.18",
]

[project.scripts]
sec-pipeline = "run:main"
